bg-remover-rembg/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
rembg==2.0.50
onnxruntime==1.16.3
Pillow==10.2.0

